// types/user-request.ts
export interface UserRequest {
  id: string;
  name: string;
  email: string;
  phone: string;
  isVerified: boolean;
  createdAt: string;
  aadharNo: string;
  pan: string;
  gender: string;
  dob: string;
  nomineeName: string;
  nomineeRelation: string;
  bankName: string;
  accountNumber: string;
  accountHolder: string;
  ifscCode: string;
  address: string;
} 